extracted most of functionality to separate file
